moves to src folder
