+++ src/commands/subscribe.ts
