+++ src/interfaces/discord.interface.ts
