+++ src/interfaces/supabase.interface.ts
