+++ src/interfaces/xbox.interface.ts
