+++ tsconfig.json
@@ -66,5 +66,6 @@
     "skipLibCheck": true,                     /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true,  /* Disallow inconsistently-cased references to the same file. */
     "resolveJsonModule": true
-  }
+  },
+  "include": ["./src/**/*"]
 }
